requirements.txt
altgraph==0.17
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
dnspython==1.16.0
docopt==0.6.2
docutils==0.16
fett==0.3.2
idna==2.10
importlib-metadata==3.7.2
macholib==1.14
networkx==2.5
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
pymongo==3.11.3
python-jsonrpc-server==0.3.4
PyYAML==5.4.1
requests==2.25.1
snooty-lextudio==1.8.10.dev0
toml==0.10.2
typing-extensions==3.7.4.3
ujson==1.35
urllib3==1.26.3
watchdog==1.0.2
zipp==3.4.1
